+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.6.0](https://github.com/Okipa/laravel-table/compare/1.5.0...1.6.0)
+
+2020-04-27
+
+* Added `Table` facade.
+* Added a cleaner architecture to handle tables:
+  * With the new table generation command, execute `php artisan make:table UsersTable --model App/Users` to generate your users table.
+  * Tables are now generated in the `app\Tables` directory.
+
 ## [1.5.0](https://github.com/Okipa/laravel-table/compare/1.4.0...1.5.0)
 
 2020-04-26